Change return value type of ILog2 to int

+++ log2.go
@@ -8,20 +8,20 @@ import (
 // ILog2 calculates log2 of a uint64 value.
 // Ported from
 // https://github.com/h2o/h2o/pull/3177/files
-func ILog2(x uint64) int64 {
-	return 63 - int64(bits.LeadingZeros64(x))
+func ILog2(x uint64) int {
+	return 63 - bits.LeadingZeros64(x)
 }
 
 // ILog2B calculates log2 of a uint64 value.
 // Ported from
 // https://twitter.com/herumi/status/1610248792254844929
-func ILog2B(x uint64) int64 {
+func ILog2B(x uint64) int {
 	f := float64(x)
 	v := math.Float64bits(f)
-	return int64(v>>52) - 1023
+	return int(v)>>52 - 1023
 }
 
-var table = []int64{
+var table = []int{
 	0, 58, 1, 59, 47, 53, 2, 60, 39, 48, 27, 54, 33, 42, 3, 61,
 	51, 37, 40, 49, 18, 28, 20, 55, 30, 34, 11, 43, 14, 22, 4, 62,
 	57, 46, 52, 38, 26, 32, 41, 50, 36, 17, 19, 29, 10, 13, 21, 56,
@@ -29,7 +29,7 @@ var table = []int64{
 }
 
 // Ported from https://stackoverflow.com/a/23000588/1391518
-func Log2ByAvernar(n uint64) int64 {
+func Log2ByAvernar(n uint64) int {
 	n |= n >> 1
 	n |= n >> 2
 	n |= n >> 4
@@ -57,16 +57,16 @@ var u8Table = []uint8{
 }
 
 // Ported from https://stackoverflow.com/a/23000588/1391518
-func Log2ByAvernarU8(n uint64) int64 {
+func Log2ByAvernarU8(n uint64) int {
 	n |= n >> 1
 	n |= n >> 2
 	n |= n >> 4
 	n |= n >> 8
 	n |= n >> 16
 	n |= n >> 32
-	return int64(u8Table[(n*0x03f6eaf2cd271461)>>58])
+	return int(u8Table[(n*0x03f6eaf2cd271461)>>58])
 }
 
-func Log2Double(n uint64) int64 {
-	return int64(math.Log2(float64(n)))
+func Log2Double(n uint64) int {
+	return int(math.Log2(float64(n)))
 }
